bugfix
page doesn't load properly past the description.

not all buttons in the admin interface change to green when saved. eg. "general content", save a campaign,

analytics: visitor per hour still not laoding

campaign management, when editing, use the whole width of the container to display the links. Move the "save" and "cancel" buttons under the list of links. Get the Banner message text to use thw whole width. of the container and move the start date field on the same row as "end date".