Add files via upload
landug page: links should open in a new window

admin > links:the "change" logo button is still too large and doesn't fit its column. Can you make it so it's not truncated? Maybe use an icon?

admin > links: there is no save button when a link or text is edited. Can you insert a message underneath a link that has been updated saying: "link has been updated" in green when changes are made

admin> campaigns: when clicking the save button under the links management, the order of links is not saved if it has been changed. The "sabe" button should record the order and which link has been disabled.

admin> campaigns: can you display the linik using the same style as the landing page, without the logo, but keep the up/down buttons.

admin> campaigns: add a date filter at the top to filter which campaigns are displayed for a certain period. Filter using the campaigns "start date"

admin> analytics: use the theme colours for the graphs. Abandoned links should use the primary colour, and the followed links / clicks per link using the secondary colour from the them.

admin> analytics: the visitors per hour doesn't display. IT should should show a timline from 0 to 24h representing the hours of the day, and a vertical stacked bar chart showing how many visits/clicks happened during that hour of the day, regardless of the day. It should folllow the date or campaign filter form the page and use the theme colour as specified in 7.

Admin> General: QR code. can you generate 2 QR codes: the first one as it is using the primary colour (no changes), The second QR code would also point to the same address, using the primary colour but includes the Logo embedded in the middle of the QR code.