Add files via upload
1. links order in campaign are still not working. The saved changes for the order of links are not reflected on the landing page or in the admin area. This is still not fixed!

2. analytics admin: move the "cumulative" text and checkbox on the same line as the date and campaign filters. 

3. admin, add the following paragraphs at the top of the page to explain how it works.
- In Analytic:
"Shows the total number of visitors to the landing page, i.e. people who have opened the landing page using the QR code or not. Abandoned links show people who did not go further than the landing page and didn't click on a link to visit a specific section. The can be considered as "abandoned" as they haven't reached the final destination. "followed links" show the actual link or links they followed. One visitor might click on multiple links. 
You can choose to see the results per day using the date filter, or per specific campaign (those will filter by campaign date automatically). If you want the total number of the period, instead of a day-by-day breakdown, click teh "cumulative" checkbox. 
The visitors by hour show the time of the day the visitor opened the page, cumulative over the days in the time period. This highlights the busier engagement period during a campaign for example."
- In Links:
This is the library of links available on this platform. You can choose the order you want them to be displayed and where they will redirect the visitor. You can hide a link from the portal, e.g. for a link that's only seasonal or temporary. You can also create campaigns to show only a specific set of links during a specified period; this way you can make the list of links relevant to the event/campaign you are participating in.