Add files via upload
1. in analytics, the graph expands vertically when using the filters. Can you make sure you keep the same height for all graph even when using the fitlers? 

2. in analytics, can you add an "export" button that will generate a PNG image of the filters and the 3 graphs in one image? The file should download on the admin's computer.

3. link management in campaings STILL DOES NOT WORK!!!! Can you review the code and fix it once and for all.

4. add a description paragraph for the Campaigns admin menu (similar to the ones you've created for analytics and links) with the following text:
"Campaigns allow you to specify a specific event that you wan to track separately, e.g. an event, exhibition or specific outreach. 
For each campaign, set the start and end dates, the name of the campaign/event, a description for you to remember, and a banner message that will be displayed on the landing page during the campaign.
You can then choose which links from the library you want to display and in which order The selected links and order will only apply during the campaign dates. If you need to add a link for that campaign, you need to create it in the Links admin menu first.
You can see the results of the campaign in this menu, or in the Analytics page, using the Campaign filter."

5. Campaigns management: enforce a rule so the "end date" cannot be set before the start date". The end date should be equal or greater to the start date"